langraph-checkpointers/requirements.txt
langgraph==1.0.7
langgraph-checkpoint-postgres==3.0.3
psycopg[binary]==3.3.2
typing-extensions==4.15.0
python-dotenv==1.2.1
google-genai>=1.60.0
fastapi==0.128.0
uvicorn==0.40.0
aiofiles==25.1.0
sse-starlette==3.2.0
python-multipart==0.0.22